+++ types.hpp
@@ -75,6 +75,6 @@ typedef std::unique_ptr<succ_vect> succ_vect_ptr;
 
 typedef const spot::option_map* const_om_ptr;
 
-typedef bool (*cut_condition_t)(const_aut_ptr, edge_t, const_om_ptr);
+typedef bool (*cut_condition_t)(spot::scc_info&, edge_t, const_om_ptr);
 
 static const state_set empty_set;