+++ ChangeLog.md
@@ -6,7 +6,7 @@ The code was almost completely rewritten.
   - `--cut-always`: jump also for all transitions leading to some accepting SCC
 * `--powerset-on-cut` starts the breakpoint/subset construction already on cut-transitions. For a marked transition s -a-> p build s -a-> (δ(s),δ_0(s),0) instead of s -a-> ({p},∅,0).
 * `--powerset-for-weak` for accepting SCC that is inherently weak in the input automaton, use only simple powerset construction in the deterministic part of the sDBA.
-* `--bscc-avoid` optimization does not include deterministic bottom SCC in the 1st component, rather jumps to the 2nd component directly.
+* `--bscc-avoid` in the 1st component, avoid deterministic SCCs with no nondeterministic successors and rather jump to the 2nd component directly.
 * `--skip-levels` enable level-skipping. For 1 edge we can "skip" multiple levels.
 * Multiple transformation types (`--via-sba`, `--via-tba`, `--via-tgba`) can be specified. The one will smallest result will be outputted.
 * `--scc-aware` enables the scc-aware optimization (default on)